refactor: move tournament configs to common/src/sports.ts
Removes duplication between frontend admin panel and backend. Both now
import TournamentConfig, TOURNAMENT_CONFIGS, and tournament objects from
common/sports. Adds dashboardPath field to TournamentConfig.
